Extract update connectors to a util method
